Add update script auto-update only

+++ AutoUpdate.sh
@@ -12,8 +12,8 @@ show_menus() {
 	echo ""	
 	echo  -e "${BLUE} U P D A T E  3 D C O I N  M A S T E R N O D E ${STD}"
 	echo ""
-	echo "1. Update Single Masternode"
-	echo "2. Update Multi Masternodes"
+	echo "1. Update Masternode & Auto-update script"
+	echo "2. Update only Auto-update script"
 	echo "3. Exit"
 	echo ""
     echo  -e "\e[1;97;41m Note!!: Update 3DCoin Core V 0.14 & Update for Auto-update   ${STD}"
@@ -24,37 +24,55 @@ read_options(){
 	local choice
 	read -p "Enter choice [ 1 - 3] " choice
 	case $choice in
-		1) Update="Update 3DCoin Masternode"
-        echo ""
-        echo  -e "${BLUE} Start ${Update}                    ${STD}"
-        rm -f /usr/local/bin/check.sh
-        rm -f /usr/local/bin/update.sh
-        rm -f /usr/local/bin/UpdateNode.sh
-        echo ""
-        cd ~
-        echo  -e "${GREEN} Get latest release                ${STD}"
-        latestrelease=$(curl --silent https://api.github.com/repos/BlockchainTechLLC/3dcoin/releases/latest | grep '"tag_name":' | sed -E 's/.*"([^"]+)".*/\1/')
-        link="https://github.com/BlockchainTechLLC/3dcoin/archive/$latestrelease.zip"
-        wget $link
-        unzip $latestrelease.zip
-        file=${latestrelease//[Vv]/3dcoin-}
-        echo ""
-        echo  -e "${GREEN} Stop Cron                         ${STD}" 
-        sudo /etc/init.d/cron stop
-        echo ""
-        echo  -e "${GREEN} Stop 3Dcoin core                  ${STD}"
-        3dcoin-cli stop
-		sleep 10
-        echo ""		
-        echo  -e "${GREEN} Make install                      ${STD}"
-        cd $file
-        ./autogen.sh
-        ./configure --disable-tests --disable-gui-tests --without-gui
-        make install-strip
-		sleep 10
-        echo ""
-        echo  -e "${GREEN} Update crontab                    ${STD}"
-        echo "#!/bin/bash
+	1) Update="Update Masternode & Auto-update"
+	    echo ""
+		echo  -e "${GREEN} Choose Type of Update                  ${STD}" 
+        PS3='Please enter your choice: '
+		options=("Update Single Masternode" "Update Multi Masternode")
+		select opt in "${options[@]}"
+		do
+			case $opt in
+				"Update Single Masternode")
+					break
+				;;
+				"Update Multi Masternode")
+					break
+				;;
+				*) echo invalid option;;
+			esac
+		done
+		#### 3Dcoin Single Node installation
+		if [ "$opt" == "Update Single Masternode" ]; then
+			echo ""
+			echo  -e "${BLUE} Start ${Update}                    ${STD}"
+			rm -f /usr/local/bin/check.sh
+			rm -f /usr/local/bin/update.sh
+			rm -f /usr/local/bin/UpdateNode.sh
+			echo ""
+			cd ~
+			echo  -e "${GREEN} Get latest release                ${STD}"
+			latestrelease=$(curl --silent https://api.github.com/repos/BlockchainTechLLC/3dcoin/releases/latest | grep '"tag_name":' | sed -E 's/.*"([^"]+)".*/\1/')
+			link="https://github.com/BlockchainTechLLC/3dcoin/archive/$latestrelease.zip"
+			wget $link
+			unzip $latestrelease.zip
+			file=${latestrelease//[Vv]/3dcoin-}
+			echo ""
+			echo  -e "${GREEN} Stop Cron                         ${STD}" 
+			sudo /etc/init.d/cron stop
+			echo ""
+			echo  -e "${GREEN} Stop 3Dcoin core                  ${STD}"
+			3dcoin-cli stop
+			sleep 10
+			echo ""		
+			echo  -e "${GREEN} Make install                      ${STD}"
+			cd $file
+			./autogen.sh
+			./configure --disable-tests --disable-gui-tests --without-gui
+			make install-strip
+			sleep 10
+			echo ""
+			echo  -e "${GREEN} Update crontab                    ${STD}"
+			echo "#!/bin/bash
 HOME=/root
 LOGNAME=root
 PATH=/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin
@@ -82,52 +100,64 @@ make install-strip
 cd ~
 rm \$latestrelease.zip
 reboot
-fi" >> /usr/local/bin/UpdateNode.sh
-        cd ~
-        cd /usr/local/bin
-        chmod 755 UpdateNode.sh
-        cd ~
-        crontab -r
-        line="@reboot /usr/local/bin/3dcoind
+fi" >> /usr/local/bin/UpdateNode.sh 
+            cd ~
+            cd /usr/local/bin
+            chmod 755 UpdateNode.sh
+            cd ~
+            crontab -r
+            line="@reboot /usr/local/bin/3dcoind
 0 0 * * * /usr/local/bin/UpdateNode.sh"
-        echo "$line" | crontab -u root -
-        echo "Crontab updated successfully"
-        echo ""
-        echo  -e "${GREEN} Start Cron                        ${STD}"
-        sudo /etc/init.d/cron start
-        echo ""		
-	    echo  -e "${GREEN} Update Finished,rebooting server  ${STD}" 
-		cd ~
-        rm $latestrelease.zip
-        reboot
-		echo ""
-        exit 0;;
+			echo "$line" | crontab -u root -
+			echo "Crontab updated successfully"
+			echo ""
+			echo  -e "${GREEN} Start Cron                        ${STD}"
+			sudo /etc/init.d/cron start
+			echo ""		
+			echo  -e "${GREEN} Update Finished,rebooting server  ${STD}" 
+			cd ~
+			rm $latestrelease.zip 
+			rm AutoUpdate.sh
+			reboot
+			echo ""
+			#### ---------------------------------------------------------- ####
+		else
+		#### 3Dcoin Multi Node updatw
+			echo ""
+			echo  -e "${BLUE} Start ${Update}                    ${STD}"
+			echo ""
+			echo  -e "${GREEN} Enter Vps ip's                    ${STD}"
+			echo ""
+			echo  -e "Please enter your vps ip's: ${RED}(Exemple: 111.111.111.111 222.222.222.222 ) ${STD}"
+			
+			unset ip
+			while [ -z ${ip} ]; do
+				read -p "IP HERE: " ip
+			done
+			
+			apt-get update
+			yes | apt-get install sshpass
+			for i in $ip
+			do
+				echo  -e "${GREEN} Connexion Vps ip $i               ${STD}"
+				echo ""
+				
+				unset rootpass
+				while [ -z ${rootpass} ]; do
+					read -s -p "Please Enter Password Root $i: " rootpass
+				done
+				
+				echo  -e "${STD}"
+				echo ""
+				echo  -e "${GREEN} SSH port Vps ip $i               ${STD}"
+				echo ""
+				
+				unset sshport
+				while [ -z ${sshport} ]; do
+					read -s -p "Please Enter ssh port $i: " sshport
+				done
 		
-        2) Update="Update 3DCoin Masternode"
-        echo ""
-        echo  -e "${BLUE} Start ${Update}                    ${STD}"
-        echo ""
-		
-        echo  -e "${GREEN} Enter Vps ip's                    ${STD}"
-		echo ""
-		echo  -e "Please enter your vps ip's: ${RED}(Exemple: 111.111.111.111 222.222.222.222 ) ${STD}"
-		unset ip
-        while [ -z ${ip} ]; do
-        read -p "IP HERE: " ip
-        done
-		apt-get update
-		yes | apt-get install sshpass
-        for i in $ip
-        do
-        echo  -e "${GREEN} Connexion Vps ip $i               ${STD}"
-        echo ""
-		unset rootpass
-        while [ -z ${rootpass} ]; do
-        read -s -p "Please Enter Password Root $i: " rootpass
-        done
-		
-		
-        sshpass -p $rootpass ssh -o StrictHostKeyChecking=no root@$i '
+        sshpass -p $rootpass ssh -p$sshport -o StrictHostKeyChecking=no root@$i '
         rm -f /usr/local/bin/check.sh
         rm -f /usr/local/bin/update.sh
         rm -f /usr/local/bin/UpdateNode.sh
@@ -210,7 +240,182 @@ fi" >> /usr/local/bin/UpdateNode.sh
         rm $latestrelease.zip
         reboot
 		echo ""'
+		rm AutoUpdate.sh
         done
+
+		fi
+        exit 0;;
+		
+        2) Update="Update only Auto-update script"
+	    echo "" 
+		echo  -e "${GREEN} Choose Type of Update                  ${STD}" 
+        PS3='Please enter your choice: '
+		options=("Update Single Masternode" "Update Multi Masternode")
+		select opt in "${options[@]}"
+		do
+			case $opt in
+				"Update Single Masternode")
+					break
+				;;
+				"Update Multi Masternode")
+					break
+				;;
+				*) echo invalid option;;
+			esac
+		done
+		#### 3Dcoin Single Node installation
+		if [ "$opt" == "Update Single Masternode" ]; then
+			echo ""
+			echo  -e "${BLUE} Start ${Update}                    ${STD}"
+			rm -f /usr/local/bin/check.sh
+			rm -f /usr/local/bin/update.sh
+			rm -f /usr/local/bin/UpdateNode.sh
+			echo ""
+			cd ~
+			echo  -e "${GREEN} Stop Cron                         ${STD}" 
+			sudo /etc/init.d/cron stop
+			echo ""
+			echo  -e "${GREEN} Update crontab                    ${STD}"
+			echo "#!/bin/bash
+HOME=/root
+LOGNAME=root
+PATH=/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin
+LANG=en_US.UTF-8
+SHELL=/bin/sh
+PWD=/root
+latestrelease=\$(curl --silent https://api.github.com/repos/BlockchainTechLLC/3dcoin/releases/latest | grep '\"tag_name\":' | sed -E 's/.*\"([^\"]+)\".*/\1/')
+localrelease=\$(3dcoin-cli -version | awk -F' ' '{print \$NF}' | cut -d \"-\" -f1)
+if [ -z \"\$latestrelease\" ] || [ \"\$latestrelease\" == \"\$localrelease\" ]; then
+exit;
+else
+cd ~
+localfile=\${localrelease//[Vv]/3dcoin-}
+rm -rf \$localfile
+link=\"https://github.com/BlockchainTechLLC/3dcoin/archive/\$latestrelease.zip\"
+wget \$link
+unzip \$latestrelease.zip
+file=\${latestrelease//[Vv]/3dcoin-}
+cd \$file
+3dcoin-cli stop
+sleep 10
+./autogen.sh
+./configure --disable-tests --disable-gui-tests --without-gui
+make install-strip
+cd ~
+rm \$latestrelease.zip
+reboot
+fi" >> /usr/local/bin/UpdateNode.sh
+            cd ~
+            cd /usr/local/bin
+            chmod 755 UpdateNode.sh
+            cd ~
+            crontab -r
+            line="@reboot /usr/local/bin/3dcoind
+0 0 * * * /usr/local/bin/UpdateNode.sh"
+			echo "$line" | crontab -u root -
+			echo "Crontab updated successfully"
+			echo ""
+			echo  -e "${GREEN} Start Cron                        ${STD}"
+			sudo /etc/init.d/cron start
+		    rm AutoUpdate.sh
+			echo ""		
+			echo  -e "${GREEN} Update Finished                   ${STD}" 
+			echo ""
+			#### ---------------------------------------------------------- ####
+		else
+		
+		
+        echo ""
+        echo  -e "${BLUE} Start ${Update}                    ${STD}"
+        echo ""
+		
+        echo  -e "${GREEN} Enter Vps ip's                    ${STD}"
+		echo ""
+		echo  -e "Please enter your vps ip's: ${RED}(Exemple: 111.111.111.111 222.222.222.222 ) ${STD}"
+		unset ip
+        while [ -z ${ip} ]; do
+        read -p "IP HERE: " ip
+        done
+		apt-get update
+		yes | apt-get install sshpass
+        for i in $ip
+        do
+        echo  -e "${GREEN} Connexion Vps ip $i               ${STD}"
+        echo ""
+		unset rootpass
+        while [ -z ${rootpass} ]; do
+        read -s -p "Please Enter Password Root $i: " rootpass
+        done
+        echo  -e "${STD}"
+        echo ""
+        echo  -e "${GREEN} SSH port Vps ip $i               ${STD}"
+        echo ""
+		unset sshport
+        while [ -z ${sshport} ]; do
+        read -s -p "Please Enter ssh port $i: " sshport
+        done
+		
+        sshpass -p $rootpass ssh -p$sshport -o StrictHostKeyChecking=no root@$i '
+			rm -f /usr/local/bin/check.sh
+			rm -f /usr/local/bin/update.sh
+			rm -f /usr/local/bin/UpdateNode.sh
+			cd ~
+			echo ""
+			echo "-----------------------------------------------------"
+			echo  -e "${GREEN} Stop Cron                         ${STD}"
+			echo "-----------------------------------------------------"
+			sudo /etc/init.d/cron stop
+			echo ""
+			echo "-----------------------------------------------------"
+			echo  -e "${GREEN} Update crontab                    ${STD}"
+			echo "-----------------------------------------------------" 
+			echo "#!/bin/bash
+HOME=/root
+LOGNAME=root
+PATH=/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin
+LANG=en_US.UTF-8
+SHELL=/bin/sh
+PWD=/root
+latestrelease=\$(curl --silent https://api.github.com/repos/BlockchainTechLLC/3dcoin/releases/latest | grep '"'"\\$"\"tag_name"\\$"\":'"' | sed -E '"'s/.*"\\$"\"([^"\\$"\"]+)"\\$"\".*/\1/'"')
+localrelease=\$(3dcoin-cli -version | awk -F'"' '"' '"'{print "\\$"\$NF}'"' | cut -d \"-\" -f1)
+if [ -z \"\$latestrelease\" ] || [ \"\$latestrelease\" == \"\$localrelease\" ]; then 
+exit;
+else
+cd ~
+localfile=\${localrelease//[Vv]/3dcoin-}
+rm -rf \$localfile
+link=\"https://github.com/BlockchainTechLLC/3dcoin/archive/\$latestrelease.zip\"
+wget \$link
+unzip \$latestrelease.zip
+file=\${latestrelease//[Vv]/3dcoin-}
+cd \$file
+3dcoin-cli stop
+sleep 10
+./autogen.sh
+./configure --disable-tests --disable-gui-tests --without-gui
+make install-strip
+cd ~
+rm \$latestrelease.zip
+reboot
+fi" >> /usr/local/bin/UpdateNode.sh
+			cd ~
+			cd /usr/local/bin
+			chmod 755 UpdateNode.sh
+			cd ~
+			crontab -r
+			line="@reboot /usr/local/bin/3dcoind
+0 0 * * * /usr/local/bin/UpdateNode.sh"
+			echo "$line" | crontab -u root -
+			echo "Crontab updated successfully"
+			sudo /etc/init.d/cron start
+			echo "" 
+			echo "-----------------------------------------------------"
+			echo  -e "${GREEN} Update Finished                   ${STD}" 
+			echo "-----------------------------------------------------"
+			echo ""'
+ 		rm AutoUpdate.sh 
+        done
+		fi
         exit 0;;
 		
 	
@@ -224,3 +429,4 @@ do
 show_menus
 read_options
 done
+
